Updated login page, Fixed delete functionality, Updated Readme, Updated Subscription Card, Verified styling
