+++ .hintrc
@@ -1,0 +1,17 @@
+{
+  "extends": [
+    "development"
+  ],
+  "hints": {
+    "compat-api/css": [
+      "default",
+      {
+        "ignore": [
+          "backface-visibility",
+          "width: fit-content",
+          "height: fit-content"
+        ]
+      }
+    ]
+  }
+}
