+++ client/components/SubscriptionCard.jsx
@@ -1,31 +1,14 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import deleteIcon from '../assets/delete.png';
+import { Context } from '../Context.js';
 
 export default function SubscriptionCard(props) {
 
   // Destructuring variables for this particular subscription card from props
   const { subscription_name, category, monthly_price } = props;
 
-  const handleDelete = (e) => {
-
-    const id = e.target.id;
-    const body = { subscriptionName: id }
-    console.log('id:', id);
-
-    fetch('/api/users', {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'Application/JSON'
-      },
-      body: JSON.stringify(body)
-    })
-      .then(res => {
-        console.log(res);
-      })
-      .catch((err) => {
-        console.log('Error', err)
-      });
-  };
+  const { handleDelete } = useContext(Context);
+  
 
   return (
     <div className="displayBox">
@@ -37,16 +20,10 @@ export default function SubscriptionCard(props) {
         <div className="displayBox-back">
           <p className="ServiceDetail"> <strong>Service Name:</strong> {subscription_name}</p>
           <p className="ServiceDetail"> <strong>Category:</strong> {category}</p>
-          <p className="ServiceDetail"> <strong>Monthly Price:</strong> {monthly_price}</p>
-          <p className="ServiceDetail"> <strong>Yearly Charge:</strong> {monthly_price * 12}</p>
-          <button className='deleteBtn' id={subscription_name} onClick={handleDelete}>Delete</button>
-        <div className="displayBox-back"> 
-          <img src={deleteIcon} className="deleteIcon" />     
-          <p className="ServiceDetail"> <strong>Service Name: </strong>{subscription_name}</p> 
-          <p className="ServiceDetail"> <strong>Category: </strong>{category}</p>
-          <p className="ServiceDetail"> <strong>Yearly Charge: </strong>${monthly_price * 12}</p>
+          <p className="ServiceDetail"> <strong>Yearly Charge:</strong> ${(monthly_price * 12).toFixed(2)}</p>
+          <img className="deleteIcon" id={subscription_name} src={deleteIcon} onClick={(e) => handleDelete(e)} />
         </div>
       </div>
     </div>
-    </div>
-  )};
+  );
+}
