+++ client/view/style.scss
@@ -357,11 +357,24 @@ label {
 }
 
 #profile-pic {
-  width: 17%;
-  height: auto;
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
   border-radius: 50%;
-  background-size: cover;
-  border: 1px solid #00061a;
+  position: relative;
+  border: .5vw solid #00061a;
+}
+
+#profile-pic-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-radius: 50%;
+  width: 17%;
+  height: 17%;
+  padding: 3px;
+  overflow: hidden;
+  background: linear-gradient(45deg, rgb(255, 170, 0), rgb(10, 205, 212) 80%);
 }
 
 #signup-box{
